+++ lib/payment_services/crypto_apis/clients/ethereum_client.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+require_relative 'base_client'
+
+class PaymentServices::CryptoApis
+  module Clients
+    class EthereumClient < PaymentServices::CryptoApis::Clients::BaseClient
+      def transaction_details(transaction_id)
+        safely_parse http_request(
+          url: "#{base_url}/txs/basic/hash/#{transaction_id}",
+          method: :GET
+        )
+      end
+
+      private
+
+      def network
+        return 'mainnet' if Rails.env.production?
+
+        currency == 'eth' ? 'rinkeby' : 'mordor'
+      end
+
+      def base_url
+        "#{API_URL}/bc/#{currency}/#{network}"
+      end
+    end
+  end
+end
